# docs/swagger.json
{
    "schemes": [
        "http",
        "ws"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Caching proxy between the Street Fighter 6 Buckler API and PostgreSQL. Serves cached battlelog data instantly and syncs with SF6 asynchronously in the background.",
        "title": "neo-shadaloo API",
        "contact": {
            "name": "Leo Alemax"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/battlelog/{userId}": {
            "get": {
                "description": "Returns the cached battlelog from the database. If the cache is stale (\u003e5 min), a background sync with the SF6 API is triggered without blocking the response. If no data exists yet, an empty replay list is returned and sync is triggered.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Get battlelog for a user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/battlelog.Battlelog"
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/calendar": {
            "get": {
                "description": "Returns battles grouped by calendar day (YYYY-MM-DD) and by weekday (0=Sun..6=Sat).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Get calendar heatmap data for a user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/battlelog.CalendarStat"
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/opponents": {
            "get": {
                "description": "Returns stats grouped by opponent character, sorted by total battles descending. Includes priority score for training recommendations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Get per-opponent-character stats for a user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/battlelog.CharStat"
                            }
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/profile": {
            "get": {
                "description": "Returns only the fighter banner info from the cached battlelog. Lighter alternative to the full battlelog endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Get profile/banner info for a user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/battlelog.FighterBannerInfo"
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "no profile data yet",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/replays": {
            "get": {
                "description": "Returns a paginated list of replays from the cached battlelog. Triggers a background sync if the cache is stale. Query params: page (default 1), limit (default 20, max 100).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Get paginated replays for a user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/battlelog.ReplayPage"
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/stats": {
            "get": {
                "description": "Returns aggregated win/loss statistics computed from the cached battlelog.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Get win/loss stats for a user",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/battlelog.WinLossStat"
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/sync": {
            "post": {
                "description": "Triggers an immediate background sync with the SF6 API for the given user, regardless of cache freshness. Returns 202 immediately; the sync runs asynchronously and notifies connected WebSocket clients when complete.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battlelog"
                ],
                "summary": "Force a battlelog sync",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "sync started",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/battlelog/{userId}/ws": {
            "get": {
                "description": "Upgrades the connection to WebSocket. The server sends a JSON message `{\"type\":\"update\",\"cachedAt\":\u003cms\u003e}` whenever a sync completes for this user. The client should re-fetch `/v1/battlelog/{userId}` on receipt. A ping is sent every 30 seconds to keep the connection alive.",
                "tags": [
                    "battlelog"
                ],
                "summary": "Subscribe to real-time battlelog updates",
                "parameters": [
                    {
                        "type": "string",
                        "example": "3378249682",
                        "description": "SF6 fighter ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols"
                    },
                    "400": {
                        "description": "userId required",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "battlelog.Battlelog": {
            "type": "object",
            "properties": {
                "bannerInfo": {
                    "$ref": "#/definitions/battlelog.FighterBannerInfo"
                },
                "cachedAt": {
                    "type": "integer"
                },
                "replays": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/battlelog.Replay"
                    }
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "battlelog.CalendarStat": {
            "type": "object",
            "properties": {
                "by_day": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/battlelog.DayStat"
                    }
                },
                "by_weekday": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/battlelog.DayStat"
                    }
                }
            }
        },
        "battlelog.CharStat": {
            "type": "object",
            "properties": {
                "clean_losses": {
                    "type": "integer"
                },
                "close_losses": {
                    "type": "integer"
                },
                "losses": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "priority_score": {
                    "type": "number"
                },
                "tool_name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "win_rate": {
                    "type": "integer"
                },
                "wins": {
                    "type": "integer"
                }
            }
        },
        "battlelog.DayStat": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "wins": {
                    "type": "integer"
                }
            }
        },
...
        "battlelog.FighterBannerInfo": {
            "type": "object",
            "properties": {
                "favorite_character_alpha": {
                    "type": "string"
                },
                "favorite_character_league_info": {
                    "$ref": "#/definitions/battlelog.FavoriteCharacterLeagueInfo"
                },
                "favorite_character_name": {
                    "type": "string"
                },
                "favorite_character_play_point": {
                    "$ref": "#/definitions/battlelog.PlayPoint"
                },
                "favorite_character_tool_name": {
                    "type": "string"
                },
                "personal_info": {
                    "$ref": "#/definitions/battlelog.PersonalInfo"
                },
                "profile_comment": {
                    "$ref": "#/definitions/battlelog.ProfileComment"
                },
                "title_data": {
                    "$ref": "#/definitions/battlelog.TitleData"
                }
            }
        },
...
        "battlelog.Replay": {
            "type": "object",
            "properties": {
                "player1_info": {
                    "$ref": "#/definitions/battlelog.PlayerInfo"
                },
                "player2_info": {
                    "$ref": "#/definitions/battlelog.PlayerInfo"
                },
                "replay_battle_sub_type_name": {
                    "type": "string"
                },
                "replay_battle_type": {
                    "type": "integer"
                },
                "replay_battle_type_name": {
                    "type": "string"
                },
                "replay_id": {
                    "type": "string"
                },
                "uploaded_at": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "battlelog.ReplayPage": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "replays": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/battlelog.Replay"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "battlelog.WinLossStat": {
            "type": "object",
            "properties": {
                "losses": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "win_pct": {
                    "type": "integer"
                },
                "wins": {
                    "type": "integer"
                }
            }
        }
    }
}